requirements.txt
pynput
pandas
numpy
websocket-client
websocket-server
tqdm